Add 64-bit support. Update dll's to v.9.05
