+++ GhostScriptSharp/GhostscriptSharp.cs
@@ -9,20 +9,6 @@ namespace GhostscriptSharp
 	/// </summary>
 	public class GhostscriptWrapper
 	{
-		#region Hooks into Ghostscript DLL
-		[DllImport("gsdll32.dll", EntryPoint = "gsapi_new_instance")]
-		private static extern int CreateAPIInstance(out IntPtr pinstance, IntPtr caller_handle);
-
-		[DllImport("gsdll32.dll", EntryPoint = "gsapi_init_with_args")]
-		private static extern int InitAPI(IntPtr instance, int argc, string[] argv);
-
-		[DllImport("gsdll32.dll", EntryPoint = "gsapi_exit")]
-		private static extern int ExitAPI(IntPtr instance);
-
-		[DllImport("gsdll32.dll", EntryPoint = "gsapi_delete_instance")]
-		private static extern void DeleteAPIInstance(IntPtr instance);
-		#endregion
-
 		#region Globals
 
 		private static readonly string[] ARGS = new string[] {
@@ -61,7 +47,10 @@ namespace GhostscriptSharp
 		/// </summary>
 		public static void GeneratePageThumbs(string inputPath, string outputPath, int firstPage, int lastPage, int width, int height)
 		{
-			CallAPI(GetArgs(inputPath, outputPath, firstPage, lastPage, width, height));
+            if (IntPtr.Size == 4)
+                API.GhostScript32.CallAPI(GetArgs(inputPath, outputPath, firstPage, lastPage, width, height));
+            else
+                API.GhostScript64.CallAPI(GetArgs(inputPath, outputPath, firstPage, lastPage, width, height));
 		}
 
 		/// <summary>
@@ -72,47 +61,10 @@ namespace GhostscriptSharp
 		/// <param name="settings">Conversion settings</param>
 		public static void GenerateOutput(string inputPath, string outputPath, GhostscriptSettings settings)
 		{
-			CallAPI(GetArgs(inputPath, outputPath, settings));
-		}
-
-		/// <summary>
-		/// Calls the Ghostscript API with a collection of arguments to be passed to it
-		/// </summary>
-		private static void CallAPI(string[] args)
-		{
-			// Get a pointer to an instance of the Ghostscript API and run the API with the current arguments
-			IntPtr gsInstancePtr;
-			lock (resourceLock)
-			{
-				CreateAPIInstance(out gsInstancePtr, IntPtr.Zero);
-				try
-				{
-					int result = InitAPI(gsInstancePtr, args.Length, args);
-
-					if (result < 0)
-					{
-						throw new ExternalException("Ghostscript conversion error", result);
-					}
-				}
-				finally
-				{
-					Cleanup(gsInstancePtr);
-				}
-			}
-		}
-
-		/// <summary>
-		/// GS can only support a single instance, so we need to bottleneck any multi-threaded systems.
-		/// </summary>
-		private static object resourceLock = new object();
-
-		/// <summary>
-		/// Frees up the memory used for the API arguments and clears the Ghostscript API instance
-		/// </summary>
-		private static void Cleanup(IntPtr gsInstancePtr)
-		{
-			ExitAPI(gsInstancePtr);
-			DeleteAPIInstance(gsInstancePtr);
+            if (IntPtr.Size == 4)
+                API.GhostScript32.CallAPI(GetArgs(inputPath, outputPath, settings));
+            else
+                API.GhostScript64.CallAPI(GetArgs(inputPath, outputPath, settings));
 		}
 
 		/// <summary>
